Update aliases
Alias the call to phpunit and phpspec in the vendor folder so users don't have to type the full path all the time (or add their own alias).

This current aliases do have the weakness of requiring being run at the project root, instead of anywhere in the project tree. Not sure how far down the "flexibility" rabbit hole we'd want to go with this, though.