Fix support for `vagrant global-status --prune`
When running the command `vagrant global-status --prune` (used by Vagrant Manager) you receive an error that `homestead.rb` can not be loaded. This appears to be because the file path is relative to where you run the vagrant command. In order to fix this issue we can pass the containing directory of the vagrant file to the `expand_path()` method when we make the default `confDir`.