Convention for supplementing nginx site configs
I'm working on a project that incorporates a large number of rewrites. I'd like to be able to version those rewrites along with the rest of the codebase, and deploy those changes in a standard way across all of my projects. So this is the concept I came up with: for Laravel-like projects in which the root of the app/site is a subfolder in the main project, a folder called `/nginx` contains the supplemental configuration files; for projects like WordPress in which the root of the site is the also the root of the project, a folder called `/.nginx` serves the same purpose. What do you think? Cheers!