Ensure Virtualbox connect the LAN cable
Somehow, Some friends and I are unable to create a working homestead with Virtualbox with just a `vagrant up`, because the LAN adapter has it's cable "disconnected" in VB settings. So we have to `vagrant up` to initialize our box, `vagrant halt` to shut it down, then go to VB settings to force the cable connection, and finally `vagrant up` to make it work. 

This simple line ensure that the cable will be connected in VB and solves our problem.