start apache before reloading it
if we try to run `service apache2 reload` with apache running, we get
the fatal error ‘apache2.service is not active, cannot reload.’ and it
kills the rest of the provisioning script.

this update first checks if apache is running. if it is, we reload like
normal.  if it isn’t, we first turn off nginx, then start apache, then
perform the reload, and finally turn nginx back on.