Updates SSL Certificate Generation
This changes a few things regarding SSL certificates and how they’re
generated. First, the code to generate a certificate has been removed
from each of the `serve-[whatever]` scripts and into it’s own script.
The certificate generation code was duplicated 6 times, and in one
instance (Symfony2) the checks to see whether or not a certificate was
already created weren’t included, like they were in the others. So this
change should help prevent things like that from happening again. DRY.

Second big change is to the way the certificates are actually generated
with OpenSSL. We’re now using an OpenSSL configuration file for each
certificate that’s generated. They’re saved along-side the
certificates. We’re doing this because the certificates must be
generated using the `subjectAltName` property to list domains, and that
property can only be added via configuration file. The `subjectAltName`
property is the primary property that is used in validating domains.
Historically, the CN (commonName) value was used to validate domains
(as a fall-back), but apparently this was never to specification and
recent browsers are removing that decade long oversight(?)

A small change worth mentioning is the removal of the CSR (Certificate
Signing Request). Since these are self-signed certificates, we don’t
need to send a request to an authority to have the certificates created.
We can just skip that step and create the certificate using the key.