Support SMB credentials in Homestead.yaml
When using SMB mounts, having to provide the SMB credentials every time the VM boots grows tiresome. This simple change adds the ability to specify those credentials in Homestead.yaml.

It goes without saying that storing these credentials in plaintext poses a security risk, but most developers are using an unprivileged, local account exclusively for this purpose, in which case any associated risk may be acceptable.