update stubs to include "features" example
this will help users get used to the new "features" syntax for optional software.

I picked the "webdriver" because it's something I imagine a relatively good portion of our users will turn on.

documentation is currently being updated as well for "features".

please confirm my JSON syntax is correct. Wasn't sure if we're supposed to use an object or an array of objects?