mark installation as "complete" at the end of script
we use an empty file in `~/.homestead-features` to mark a "feature" as installed. we should move to the end of the file, so we know the install has fully completed successfully.

I ran into this issue with the "webdriver" install. because Chrome does not have an ARM version, the install would fail, but the file would already be created, the provision scipt thought it succeeded.